Buscar por ubicacion cliente fixes (#3)
* componente LocationAutocomplete mas flexible y borrado archivos innecesarios
