+++ src/components/App.jsx
@@ -12,7 +12,7 @@ const App = () => {
             <Bar/>
             <Container maxWidth="lg" >
                 <Switch>
-                  <Route exact path="/" render={LoggedOutHomePage}/>
+                  <Route exact path="/" component={LoggedOutHomePage}/>
                 </Switch>
             </Container>
         </BrowserRouter>
